Use docker node:10-alpine instead of node:lts-alpine.
lts-alpine is a mutable tag. Docker Hub will continue to update the image, so it no longer points to the node version when the project was created, but to the current LTS latest version.

In the current LTS latest version 22.17, siteserver will throw the error:0308010C:digital envelope routines::unsupported.

Refs: https://stackoverflow.com/questions/69692842/error-message-error0308010cdigital-envelope-routinesunsupported